requirements.txt
pandas>=2.2.3,<3.0.0
numpy>=1.26.4,<3.0.0
PyYAML>=6.0.2,<7.0.0
pyspark>=3.5.2,<4.0.0
duckdb>=1.1.3,<2.0.0
faiss-cpu>=1.8.0.post1,<2.0.0
scikit-learn>=1.5.2,<2.0.0
langgraph>=0.2.56,<1.0.0
streamlit>=1.39.0,<2.0.0
plotly>=5.24.1,<7.0.0
mlflow>=2.17.0,<4.0.0
python-dotenv>=1.0.1,<2.0.0
pytest>=8.3.3,<9.0.0
faiss-cpu
